+++ src/game/themes/summer.ts
@@ -1,0 +1,39 @@
+import { BlockColor } from "../types";
+
+export const shapeMap: Record<BlockColor, string> = {
+  black: "☀️",
+  brown: "🕶️",
+  darkgreen: "️🌴",
+  yellow: "🦩",
+  aqua: "️🌻",
+  pink: "🦋",
+  purple: "️️️🐬",
+  blue: "🍦",
+  red: "🍨",
+  white: "️️💃",
+  green: "️🕺",
+  darkblue: "️️🎵",
+  turquoise: "🎀",
+  orange: "🫧",
+  lightyellow: "️️☁️",
+  gray: "️🧩"
+};
+
+export const colorMap: Record<BlockColor, string> = {
+  red: "#00FA9A",
+  white: "#FF4500",
+  yellow: "#FF6F61",
+  blue: "#ADFF2F",
+  purple: "#718096",
+  black: "#FFD700",
+  green: "#FFDAB9",
+  darkgreen: "#32CD32",
+  darkblue: "#20B2AA",
+  aqua: "#87CEEB",
+  brown: "#FFFACD",
+  pink: "#FF69B4",
+  turquoise: "#4682B4",
+  orange: "#E0FFFF",
+  lightyellow: "#00BFFF",
+  gray: "#FFA07A"
+};